moves abstract teacher out of agents module
